Created the blockerGUI python
Has a list of blocked apps that can be changed that will close
applications. Will check if has admin permission. Then checks what
applications are running. If an app that is running is on the block
list, it will close that app and will close it if the user tries to
open the application again. There is also a dry run function in which
the user can run the application safely without actually closing
anything and it will show what the blocker would close. Rest is GUI
stuff that is not required for the program.